TEST / set up so can reset hard back to version without compiled files everywhere
