+++ .gitignore
@@ -1,4 +1,4 @@
 node_modules
 .env
-dist
+/dist
 server/db/config/postgresConfig.json
