+++ package.json
@@ -5,6 +5,7 @@
   "main": "index.jsx",
   "scripts": {
     "build-prod": "webpack --config webpack.config.js",
+    "compile-test": "tsc --project ./test",
     "test": "npx sequelize db:drop --env test && npx sequelize db:create --env test && mocha",
     "build": "webpack --config webpack.config.js --watch",
     "start": "nodemon server/index.ts",
@@ -27,7 +28,6 @@
     "@babel/preset-react": "^7.23.3",
     "@mapbox/search-js-react": "^1.0.0-beta.18",
     "@react-icons/all-files": "^4.1.0",
-    "@types/jest": "^29.5.10",
     "@types/multer": "^1.4.11",
     "@types/react": "^18.2.41",
     "@types/react-dom": "^18.2.17",
@@ -62,7 +62,9 @@
     "@babel/core": "^7.23.5",
     "@babel/preset-env": "^7.23.5",
     "@types/bootstrap": "^5.2.10",
+    "@types/chai": "^4.3.12",
     "@types/express": "^4.17.21",
+    "@types/mocha": "^10.0.6",
     "@types/node": "^20.10.3",
     "@types/react-datepicker": "^4.19.5",
     "@types/react-modal": "^3.16.3",
@@ -88,7 +90,7 @@
     "sass-loader": "^13.3.2",
     "style-loader": "^3.3.3",
     "ts-loader": "^9.5.1",
-    "ts-node": "^10.9.1",
+    "ts-node": "^10.9.2",
     "typescript": "^5.3.2",
     "webpack": "^5.89.0",
     "webpack-bundle-analyzer": "^4.10.1",
