+++ test/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "include": ["./index.test.ts"],
+  "exclude": ["../server/**/*", "../client/**/*", "../dist"],
+  "compilerOptions": {
+    "outDir": "./dist",
+    "rootDir": "./",
+  },
+}
