+++ tsconfig.json
@@ -1,6 +1,7 @@
 {
   "compilerOptions": {
-    "outDir": "./dist/",
+    "rootDir": "./" ,
+    "outDir": "./dist",
     "noImplicitAny": true,
     "module": "commonjs",
     "target": "es6",
@@ -14,5 +15,5 @@
     "emitDecoratorMetadata": true
     // "noEmit": true,
     // "allowImportingTsExtensions": true,
-  }
+  },
 }
